requirements.txt
fastapi==0.100.1
python-dotenv==1.0.0
SQLAlchemy==1.4.41
SQLAlchemy-Utils==0.38.3
httpx==0.25.0
uvicorn==0.23.1
gunicorn
asyncpg